# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CBDC Clearing Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Node {
  id: number;
  currency: string;
  name: string;
  credibility: number;
  gdp: number;
  amlRisk: number;
  dailyVolCapM: number;
  status: string;
  /** @nullable */
  color?: string | null;
}

export interface Corridor {
  id: number;
  srcCurrency: string;
  tgtCurrency: string;
  fxRate: number;
  costBps: number;
  latencyS: number;
  liquidityM: number;
  friction: number;
  pvp: boolean;
  status?: string;
}

export interface Transaction {
  id: number;
  /** @nullable */
  txHash?: string | null;
  srcCurrency: string;
  tgtCurrency: string;
  amount: number;
  /** @nullable */
  amountConverted?: number | null;
  /** @nullable */
  purposeCode?: string | null;
  /** @nullable */
  entityType?: string | null;
  status: string;
  /** @nullable */
  route?: string | null;
  /** @nullable */
  costBps?: number | null;
  /** @nullable */
  latencyS?: number | null;
  /** @nullable */
  fxRiskScore?: number | null;
  /** @nullable */
  complianceScore?: number | null;
  /** @nullable */
  compositeScore?: number | null;
  /** @nullable */
  blockHash?: string | null;
  /** @nullable */
  prevBlockHash?: string | null;
  createdAt: string;
  /** @nullable */
  settledAt?: string | null;
}

export interface TransactionInput {
  srcCurrency: string;
  tgtCurrency: string;
  amount: number;
  purposeCode?: string;
  entityType?: string;
  alphaWeight?: number;
  betaWeight?: number;
  gammaWeight?: number;
  deltaWeight?: number;
}

export interface RouteRequest {
  srcCurrency: string;
  tgtCurrency: string;
  amount: number;
  purposeCode?: string;
  entityType?: string;
  alphaWeight?: number;
  betaWeight?: number;
  gammaWeight?: number;
  deltaWeight?: number;
}

export interface AlternativeRoute {
  path: string[];
  latencyS: number;
  liquidityM: number;
  totalFeesBps: number;
  status: string;
  compositeScore?: number;
}

export interface RouteResult {
  optimalRoute: string[];
  compositeScore: number;
  costBps: number;
  latencyS: number;
  fxRiskScore: number;
  complianceScore: number;
  alternativeRoutes: AlternativeRoute[];
}

export interface FxRate {
  id: number;
  srcCurrency: string;
  tgtCurrency: string;
  rate: number;
  updatedAt: string;
}

export interface FxRisk {
  srcCurrency: string;
  tgtCurrency: string;
  riskScore: number;
  ewmaVol: number;
  ar1Forecast: number;
  regime: string;
}

export interface NodeLiquidity {
  id: number;
  currency: string;
  totalLiquidityB: number;
  availableLiquidityB: number;
  reservedLiquidityB: number;
  utilizationPct: number;
  efficiencyScore?: number;
  status: string;
}

export interface LiquidityInjection {
  currency: string;
  amountB: number;
}

export interface StressTestInput {
  scenario: string;
  srcNode?: string;
  dstGroup?: string;
  trials: number;
  varPct?: number;
  includeLiquidityContagion?: boolean;
}

export interface StressTestResult {
  id: number;
  scenario: string;
  trials?: number;
  successRate: number;
  varCost: number;
  avgLatencyS: number;
  resilienceScore: string;
  /** @nullable */
  baselineSuccessRate?: number | null;
  /** @nullable */
  baselineVarCost?: number | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalTransactions: number;
  successRate: number;
  totalVolumeM: number;
  avgCostBps: number;
  avgLatencyS: number;
  activeCorridors: number;
  activeNodes: number;
  totalCost: number;
  nodeLoad?: number;
  queueDepthMs?: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  message: string;
  srcCurrency: string;
  tgtCurrency: string;
  amount: number;
  status: string;
  timestamp: string;
  /** @nullable */
  txHash?: string | null;
}

export type ListTransactionsParams = {
limit?: number;
status?: string;
};

export type ListActivityParams = {
limit?: number;
};