# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CBDC Clearing Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Initiate a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/settle:
    post:
      operationId: settleTransaction
      tags: [transactions]
      summary: Execute settlement for a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settlement result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /route/optimize:
    post:
      operationId: optimizeRoute
      tags: [transactions]
      summary: Compute optimal route for a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Route optimization result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteResult"
...
components:
  schemas:
...
    Transaction:
      type: object
      required: [id, srcCurrency, tgtCurrency, amount, status, createdAt]
      properties:
        id:
          type: integer
        txHash:
          type: ["string", "null"]
        srcCurrency:
          type: string
        tgtCurrency:
          type: string
        amount:
          type: number
        amountConverted:
          type: ["number", "null"]
        purposeCode:
          type: ["string", "null"]
        entityType:
          type: ["string", "null"]
        status:
          type: string
        route:
          type: ["string", "null"]
        costBps:
          type: ["number", "null"]
        latencyS:
          type: ["number", "null"]
        fxRiskScore:
          type: ["number", "null"]
        complianceScore:
          type: ["number", "null"]
        compositeScore:
          type: ["number", "null"]
        blockHash:
          type: ["string", "null"]
        prevBlockHash:
          type: ["string", "null"]
        createdAt:
          type: string
        settledAt:
          type: ["string", "null"]
    TransactionInput:
      type: object
      required: [srcCurrency, tgtCurrency, amount]
      properties:
        srcCurrency:
          type: string
        tgtCurrency:
          type: string
        amount:
          type: number
        purposeCode:
          type: string
        entityType:
          type: string
        alphaWeight:
          type: number
        betaWeight:
          type: number
        gammaWeight:
          type: number
        deltaWeight:
          type: number
    RouteRequest:
      type: object
      required: [srcCurrency, tgtCurrency, amount]
      properties:
        srcCurrency:
          type: string
        tgtCurrency:
          type: string
        amount:
          type: number
        purposeCode:
          type: string
        entityType:
          type: string
        alphaWeight:
          type: number
        betaWeight:
          type: number
        gammaWeight:
          type: number
        deltaWeight:
          type: number
    RouteResult:
      type: object
      required: [optimalRoute, compositeScore, costBps, latencyS, fxRiskScore, complianceScore, alternativeRoutes]
      properties:
        optimalRoute:
          type: array
          items:
            type: string
        compositeScore:
          type: number
        costBps:
          type: number
        latencyS:
          type: number
        fxRiskScore:
          type: number
        complianceScore:
          type: number
        alternativeRoutes:
          type: array
          items:
            $ref: "#/components/schemas/AlternativeRoute"
    AlternativeRoute:
      type: object
      required: [path, latencyS, liquidityM, totalFeesBps, status]
      properties:
        path:
          type: array
          items:
            type: string
        latencyS:
          type: number
        liquidityM:
          type: number
        totalFeesBps:
          type: number
        status:
          type: string
        compositeScore:
          type: number
...