# lib/api-zod/src/generated/types/alternativeRoute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CBDC Clearing Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface AlternativeRoute {
  path: string[];
  latencyS: number;
  liquidityM: number;
  totalFeesBps: number;
  status: string;
  compositeScore?: number;
}