# lib/api-zod/src/generated/types/routeRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CBDC Clearing Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteRequest {
  srcCurrency: string;
  tgtCurrency: string;
  amount: number;
  purposeCode?: string;
  entityType?: string;
  alphaWeight?: number;
  betaWeight?: number;
  gammaWeight?: number;
  deltaWeight?: number;
}