# lib/api-zod/src/generated/types/routeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CBDC Clearing Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { AlternativeRoute } from './alternativeRoute';

export interface RouteResult {
  optimalRoute: string[];
  compositeScore: number;
  costBps: number;
  latencyS: number;
  fxRiskScore: number;
  complianceScore: number;
  alternativeRoutes: AlternativeRoute[];
}